BigDecimal instead of all long/double types

+++ loans-services/src/main/java/com/example/loansservices/dto/LoansDto.java
@@ -4,6 +4,7 @@ import com.example.loansservices.model.Loans;
 import jakarta.persistence.Column;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -19,13 +20,13 @@ public class LoansDto implements   Dto{
     private LocalDate endDt;
     private LocalDate startDt;
     private Loans.LoanType loanType;
-    private Long totalLoan;
+    private BigDecimal totalLoan;
     private int loanTenureInYears;
-    private Long amountPaid;
-    private Long emiAmount;
+    private BigDecimal amountPaid;
+    private BigDecimal emiAmount;
     private Double Rate_of_Interest;
     private int totalInstallmentsInNumber;
     private int installmentsPaidInNumber;
     private int installmentsRemainingInNumber;
-    private Long outstandingAmount;
+    private BigDecimal outstandingAmount;
 }
